refactor(domain): 記事ドメインをモジュール分割して整理
- article.rsをmodel, repository, serviceに分割
- 公開APIをmod.rsで再エクスポート
- テストケースを各モジュールに移動
